requirements.txt
requests
websocket-client
Flask
python-dotenv
py-sversion